backend/emotion-detector/requirements.txt
fastapi
uvicorn
opencv-python-headless
numpy
hsemotion
facenet-pytorch
supabase
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
